Rounding gain

+++ cogs/profile.py
@@ -119,7 +119,7 @@ class Profile(commands.Cog):
         """Récupérer des Lulux Coins chaque heure."""
         earned = random.randint(100, 300)
         if any([d for d in interaction.user.roles if d.is_premium_subscriber() is True]):
-            earned = earned * 1.15
+            earned = int(earned * 1.15)
 
         await self.client.mongo.update_user_data_document(interaction.user.id, {"$inc": {"bank": earned}})
 
@@ -134,7 +134,7 @@ class Profile(commands.Cog):
         """Récupérer des Lulux Coins chaque jour."""
         earned = random.randint(2000, 3000)
         if any([d for d in interaction.user.roles if d.is_premium_subscriber() is True]):
-            earned = earned * 1.15
+            earned = int(earned * 1.15)
 
         await self.client.mongo.update_user_data_document(interaction.user.id, {"$inc": {"bank": earned}})
